+++ cluster-network.tf
@@ -33,10 +33,10 @@ resource "oci_core_cluster_network" "cluster_network" {
     size                      = var.node_count
     display_name              = local.cluster_name
   }
+  defined_tags = local.user_cost_tags
   freeform_tags = {
       "cluster_name" = local.cluster_name
       "parent_cluster" = local.cluster_name
-      "user" = var.tags
   }
   placement_configuration {
     availability_domain = var.ad