+++ slurm_ha.tf
@@ -58,6 +58,7 @@ resource "null_resource" "backup" {
   depends_on = [oci_core_instance.backup] 
   triggers = merge({
     slurm_user_tags_enabled = tostring(var.slurm && var.slurm_user_tags_enabled)
+    user_cost_tag_key       = local.user_cost_tag_key
     backup = oci_core_instance.backup[0].id
   }, var.slurm_job_notifications_enabled ? {
     slurm_notification_payload = local.slurm_notification_payload_hash
@@ -193,6 +194,7 @@ resource "null_resource" "cluster_backup" {
   ]
   triggers = merge({
     slurm_user_tags_enabled = tostring(var.slurm && var.slurm_user_tags_enabled)
+    user_cost_tag_key       = local.user_cost_tag_key
     cluster_instances           = join(", ", local.cluster_instances_names)
     cluster_instance_ids        = join(", ", local.cluster_instances_ids)
     local_block_volume_settings = sha256(jsonencode([var.use_local_block_volume, var.local_block_volume_size, var.local_block_volume_performance, var.local_block_volume_mount_point]))